# Speed up when looping on many single lines